--- src/main/java/com/example/controller/phaseOne/PhaseOneController.java
@@ -1,10 +1,0 @@
-package com.example.controller.phaseOne;
-
-public interface PhaseOneController {
-
-  void PresentPhase();
-
-  long GenPrime(int bit, String fileName) throws Exception;
-
-  long[] GenRandomNoWithInverse(long prime);
-}
