+++ src/main/java/com/example/service/exponentiation/ExponentiationService.java
@@ -1,7 +1,5 @@
 package com.example.service.exponentiation;
 
-import java.math.BigInteger;
-
 public interface ExponentiationService {
 
   long bruteForceExpo(long base, long expo);
@@ -9,6 +7,4 @@ public interface ExponentiationService {
   long fastExpo(long base, long expo);
 
   long fastExpo(long base, long expo, long mod);
-
-  BigInteger binaryFastExpoReturnBigInt(long base, long expo);
 }
