+++ src/main/java/com/example/service/exponentiation/ExponentiationServiceImpl.java
@@ -1,6 +1,5 @@
 package com.example.service.exponentiation;
 
-import java.math.BigInteger;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -55,23 +54,6 @@ public class ExponentiationServiceImpl implements ExponentiationService {
     }
 
     //        System.out.println("base: " + base + " expo: " + expo + " result: " + result);
-    return result;
-  }
-
-  @Override
-  public BigInteger binaryFastExpoReturnBigInt(long base, long expo) {
-    BigInteger result = BigInteger.ONE;
-    BigInteger a = BigInteger.valueOf(base);
-    long b = expo;
-
-    while (b > 0) {
-      if ((b & 1) == 1) {
-        result = result.multiply(a);
-      }
-      a = a.multiply(a);
-      b >>= 1;
-      System.out.println("b: " + b);
-    }
     return result;
   }
 }
