--- src/main/java/com/example/service/generator/GeneratorService.java
@@ -1,6 +1,0 @@
-package com.example.service.generator;
-
-public interface GeneratorService {
-
-  long findGenerator(long prime);
-}
